updated readme.md and added test coverage
